pyproject.toml
[tool.poetry]
name = "mojob-test"
version = "0.1.0"
description = ""
authors = ["Vaipik <n.khylko@gmail.com>"]
readme = "README.md"
packages = [{include = "mojob_test"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"
pyyaml = "^6.0"
uritemplate = "^4.1.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.13.2"
black = "^23.3.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
